Moved GeometrySerde to serde folder Created abstract GeometrySerde class for specific serde implementations Created ShapeGeometrySerde and WKBGeometrySerde that implements GeometrySerde Added tests for WKB Serde
